Add files via upload
In this assignment, we familiarize ourselves with modern machine learning; in particular deep learning. Image categorization is probably the most studied application example of deep learning. As an example of designing a deep neural network, we will consider smile detection using a deep network and deploy that for real-time execution. For training the network, we use the GENKI-4k Face Expression and Pose Dataset [1], where the task is to categorize images of faces into two classes based on the facial expression; smile and non-smile.